+++ CHANGELOG.md
@@ -3,6 +3,23 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0]
+
+### Added
+
+* `struct-field-names-as-array-derive` crate
+
+* `FieldNamesAsArray` trait
+
+* `FieldNamesAsSlice` trait
+
+* `FieldNamesAsSlice` procedural macro
+
+### Changed
+
+* upgraded rust edition from 2018 to 2021
+
+* `struct-field-names-as-array-derive`: `syn v1 -> v2`
 
 ## [0.2.0]